requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
python-multipart
httpx
chromadb
google-generativeai
pypdf2